pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-bom"
dynamic = ["version"]
description = "AI Bill of Materials — discover and inventory all AI/LLM agents, models, and API integrations across your infrastructure."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Trusera", email = "info@trusera.dev" },
]
keywords = ["ai", "bom", "security", "llm", "agents", "compliance", "n8n", "shadow-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "tomli>=2.0.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
ai-bom = "ai_bom.cli:app"

[project.urls]
Homepage = "https://trusera.dev"
Repository = "https://github.com/trusera/ai-bom"
Issues = "https://github.com/trusera/ai-bom/issues"
Changelog = "https://github.com/trusera/ai-bom/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "src/ai_bom/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
    ".venv",
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".github",
    "tests",
    "htmlcov",
    ".coverage",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_bom"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
